learning/src/test/java/io/bitsmart/bdd/learning/junit5/watcher/SuccessTestWatcher.java
/*
 * Smart BDD - The smart way to do behavior-driven development.
 * Copyright (C)  2021  James Bayliss
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 *
 * You should have received a copy of the GNU General Public License
 * along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */

package io.bitsmart.bdd.learning.junit5.watcher;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.api.extension.TestWatcher;

import java.util.ArrayList;
import java.util.List;

public class SuccessTestWatcher implements TestWatcher {
    private static List<String> results = new ArrayList<>();

    @Override
    public void testSuccessful(ExtensionContext context) {
        results.add("success");
    }

    public static List<String> getResults() {
        return results;
    }

    public static void clearResults() {
        results.clear();
    }
}
